schemachanger/scdeps/sctestdeps: decompose constructor
Release note: None
