+++ pkg/sql/schemachanger/scbuild/builder_test.go
@@ -29,6 +29,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/screl"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
+	"github.com/cockroachdb/cockroach/pkg/sql/sessiondatapb"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
@@ -56,7 +58,21 @@ func TestBuilderAlterTable(t *testing.T) {
 			{
 				"test_dependencies",
 				func(t *testing.T, s serverutils.TestServerInterface, tdb *sqlutils.SQLRunner, fn func(scbuild.Dependencies)) {
-					fn(sctestdeps.NewTestDependencies(ctx, t, tdb, nil /* testingKnobs */, nil /* statements */))
+					// Create test dependencies and execute the schema changer.
+					// The schema changer test dependencies do not hold any reference to the
+					// test cluster, here the SQLRunner is only used to populate the mocked
+					// catalog state.
+					fn(sctestdeps.NewTestDependencies(
+						sctestdeps.WithDescriptors(sctestdeps.ReadDescriptorsFromDB(ctx, t, tdb)),
+						sctestdeps.WithNamespace(sctestdeps.ReadNamespaceFromDB(t, tdb)),
+						sctestdeps.WithCurrentDatabase(sctestdeps.ReadCurrentDatabaseFromDB(t, tdb)),
+						sctestdeps.WithSessionData(sctestdeps.ReadSessionDataFromDB(t, tdb, func(
+							sd *sessiondata.SessionData,
+						) {
+							// For setting up a builder inside tests we will ensure that the new schema
+							// changer will allow non-fully implemented operations.
+							sd.NewSchemaChangerMode = sessiondatapb.UseNewSchemaChangerUnsafe
+						}))))
 				},
 			},
 		} {
