+++ pkg/sql/schemachanger/scdeps/sctestdeps/BUILD.bazel
@@ -3,6 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "sctestdeps",
     srcs = [
+        "config.go",
+        "database_state.go",
         "test_deps.go",
         "test_state.go",
     ],
